+++ consensus/src/dag/adapter.rs
@@ -6,6 +6,7 @@ use super::{
     observability::counters::{NUM_NODES_PER_BLOCK, NUM_ROUNDS_PER_BLOCK},
 };
 use crate::{
+    block_storage::tracing::{observe_block, BlockStage},
     consensusdb::{CertifiedNodeSchema, ConsensusDB, DagVoteSchema, NodeSchema},
     counters::update_counters_for_committed_blocks,
     dag::{
@@ -165,9 +166,8 @@ impl OrderedNotifier for OrderedNotifierAdapter {
 
         NUM_NODES_PER_BLOCK.observe(ordered_nodes.len() as f64);
         let rounds_between = {
-            let anchor_node = ordered_nodes.first().map_or(0, |node| node.round());
-            let lowest_round_node = ordered_nodes.last().map_or(0, |node| node.round());
-            anchor_node.saturating_sub(lowest_round_node)
+            let lowest_round_node = ordered_nodes.first().map_or(0, |node| node.round());
+            round.saturating_sub(lowest_round_node)
         };
         NUM_ROUNDS_PER_BLOCK.observe((rounds_between + 1) as f64);
 
@@ -196,6 +196,9 @@ impl OrderedNotifier for OrderedNotifierAdapter {
             .write()
             .insert(block_info.round(), Instant::now());
         let block_created_ts = self.block_ordered_ts.clone();
+
+        observe_block(block.block().timestamp_usecs(), BlockStage::ORDERED);
+
         let blocks_to_send = OrderedBlocks {
             ordered_blocks: vec![block],
             ordered_proof: LedgerInfoWithSignatures::new(