+++ network/framework/src/protocols/wire/messaging/v1/test.rs
@@ -239,7 +239,6 @@ proptest! {
         let (mut msg_tx, msg_rx) = aptos_channels::new_test(1024);
         let mut outbound_stream = OutboundStream::new(128, 64 * 255, stream_tx);
         let mut inbound_stream = InboundStreamBuffer::new(255);
-
         let messages_clone = messages.clone();
         let f_stream_all = async move {
             for message in messages_clone {